+++ src/components/About/About.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import Card from '@material-ui/core/Card';
+import CardContent from '@material-ui/core/CardContent';
+import Typography from '@material-ui/core/Typography';
+
+const About = () => (
+  <Card>
+    <CardContent>
+      <Typography variant="h5">
+        About
+      </Typography>
+    </CardContent>
+  </Card>
+);
+
+export default About;
