+++ src/components/Contacts/Contacts.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import Card from '@material-ui/core/Card';
+import CardContent from '@material-ui/core/CardContent';
+import Typography from '@material-ui/core/Typography';
+
+const Contacts = () => (
+  <Card>
+    <CardContent>
+      <Typography variant="h5">
+        Contacts
+      </Typography>
+    </CardContent>
+  </Card>
+);
+
+export default Contacts;
